Added the ability to pull with a certain force. Algorithm changes.
Rather than the actuator retreating to a certain position when a command signal is recieved, the actuator now pulls with ~20lbs of force for a certain duration.
Trials setting a 10kg weight on the actuator showed that the previous algorithm for oscilation-arrest caused a different sort of oscilation in this new setting.
Major changes to the algorithm were needed to find a compromise between the two scenarios, but although oscillation arrest is now worse, I am now less worried about unexpected extreme oscilation cases.
It's a bit of a mess, and at this point it's complicated enough that I'm actually not sure which parts of the algorithm are actually productive. More work is needed.